+++ src/main/java/com/kdkj/caijin/service/impl/InformationServiceImpl.java
@@ -4,6 +4,7 @@ import com.kdkj.caijin.dao.InformationDao;
 import com.kdkj.caijin.entity.Information;
 import com.kdkj.caijin.service.InformationService;
 import com.kdkj.caijin.util.CopyObj;
+import com.kdkj.caijin.util.ErrMsgException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -13,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -42,7 +44,11 @@ public class InformationServiceImpl implements InformationService {
     @Override
     public int update(Information information) throws IllegalAccessException, InstantiationException {
         if (information != null) {
-            Information oldInformation = informationDao.findById(information.getId()).get();
+            Optional<Information> byId = informationDao.findById(information.getId());
+            if (!byId.isPresent()) {
+                throw new ErrMsgException("不存在该id");
+            }
+            Information oldInformation = byId.get();
             CopyObj.copyObjNotNullFieldsAsObj(information, oldInformation);
             return 1;
         }
@@ -57,6 +63,10 @@ public class InformationServiceImpl implements InformationService {
     @Override
     public int deleteById(String id) {
         if (!StringUtils.isEmpty(id)) {
+            Optional<Information> byId = informationDao.findById(id);
+            if (!byId.isPresent()) {
+                throw new ErrMsgException("不存在该id");
+            }
             informationDao.deleteById(id);
             return 1;
         }
@@ -70,5 +80,16 @@ public class InformationServiceImpl implements InformationService {
             return 1;
         }
         return 0;
+    }
+
+    @Override
+    public List<Information> findBySourceNotNull() {
+        return informationDao.findBySourceNotNull();
+    }
+
+    @Override
+    public int insertAll(List<Information> list) {
+        informationDao.saveAll(list);
+        return 1;
     }
 }
