+++ src/main/java/com/kdkj/caijin/service/impl/OpinionServiceImpl.java
@@ -1,7 +1,10 @@
 package com.kdkj.caijin.service.impl;
 
 import com.kdkj.caijin.dao.OpinionDao;
+import com.kdkj.caijin.dao.UsersDao;
 import com.kdkj.caijin.entity.Opinion;
+import com.kdkj.caijin.entity.Users;
+import com.kdkj.caijin.enums.OpinionInfo;
 import com.kdkj.caijin.service.OpinionService;
 import com.kdkj.caijin.util.CopyObj;
 import com.kdkj.caijin.util.ErrMsgException;
@@ -13,24 +16,36 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
+import java.util.Optional;
+
 @Service
 @Transactional
 public class OpinionServiceImpl implements OpinionService {
     @Autowired
     private OpinionDao opinionDao;
-
+    @Autowired
+    private UsersDao usersDao;
     @Override
     public Page<Opinion> findAll(PageRequest pageRequest) {
         return opinionDao.findAll(pageRequest);
     }
 
     @Override
-    public int insert(Opinion opinion) {
+    public Opinion insert(Opinion opinion) {
         if (opinion != null) {
-            opinionDao.save(opinion);
-            return 1;
+            if (StringUtils.isEmpty(opinion.getUserid())) {
+                throw new ErrMsgException("用户id不能为空");
+            }
+            Optional<Users> byId = usersDao.findById(opinion.getUserid());
+            if (!byId.isPresent()) {
+                throw new ErrMsgException("没有该id的用户");
+            }
+            opinion.setAnswer(OpinionInfo.NOT_ANSWER.getCode());
+            opinion.setSee(OpinionInfo.NOT_SEE.getCode());
+            Opinion save = opinionDao.save(opinion);
+            return save;
         }
-        return 0;
+        return null;
     }
 
     @Override
@@ -55,6 +70,10 @@ public class OpinionServiceImpl implements OpinionService {
     @Override
     public int deleteById(String id) {
         if (!StringUtils.isEmpty(id)) {
+            Optional<Opinion> byId = opinionDao.findById(id);
+            if (!byId.isPresent()) {
+                throw new ErrMsgException("没有该id");
+            }
             opinionDao.deleteById(id);
             return 1;
         }
